AirdropSellBot/requirements.txt
# Библиотеки для работы с API бирж
python-binance==1.0.19    # Для Binance API
python-okx==0.3.5        # Для OKX API (последняя версия на момент обновления)
pybit==5.5.0             # Для Bybit API (универсальный клиент)
okx-api==1.0.5

# Утилиты для улучшения кода
loguru==0.7.2            # Простое и мощное логирование вместо print
python-dotenv==1.0.1     # Для загрузки API-ключей из .env файла
okx~=2.1.2
binance~=0.3
requests~=2.32.3
urllib3~=1.26.12